Move functions to a new polyfill.ts
