+++ packages/core/src/tools/utils/polyfills.spec.ts
@@ -1,4 +1,4 @@
-import { startsWith, arrayFrom, cssEscape, elementMatches } from './polyfills'
+import { startsWith, arrayFrom } from './polyfills'
 
 describe('polyfills', () => {
   describe('startWith', () => {
@@ -24,29 +24,6 @@ describe('polyfills', () => {
       div.classList.add('foo')
 
       expect(arrayFrom(div.classList)).toEqual(['foo'])
-    })
-  })
-
-  describe('cssEscape', () => {
-    it('should escape a string', () => {
-      expect(cssEscape('.foo#bar')).toEqual('\\.foo\\#bar')
-      expect(cssEscape('()[]{}')).toEqual('\\(\\)\\[\\]\\{\\}')
-      expect(cssEscape('--a')).toEqual('--a')
-      expect(cssEscape('\0')).toEqual('\ufffd')
-    })
-  })
-
-  describe('elementMatches', () => {
-    it('should return true if the element matches the selector', () => {
-      const element = document.createElement('div')
-      element.classList.add('foo')
-      expect(elementMatches(element, '.foo')).toEqual(true)
-    })
-
-    it('should return false if the element does not match the selector', () => {
-      const element = document.createElement('div')
-      element.classList.add('bar')
-      expect(elementMatches(element, '.foo')).toEqual(false)
     })
   })
 })
