+++ packages/rum-core/src/browser/htmlDomUtils.spec.ts
@@ -9,7 +9,6 @@ import {
   isNodeShadowHost,
   forEachChildNodes,
   hasChildNodes,
-  getClassList,
 } from './htmlDomUtils'
 
 describe('isTextNode', () => {
@@ -219,37 +218,3 @@ if (!isIE()) {
     })
   })
 }
-
-describe('getParentElement', () => {
-  it('should return the classList of an element that supports the classList attribute', () => {
-    const divElement = appendElement('<div class="foo bar"></div>')
-    const classList = getClassList(divElement)
-    expect(classList[0]).toEqual('foo')
-    expect(classList[1]).toEqual('bar')
-  })
-
-  it('should return the classList of an element that does not support the classList attribute (e.g.: svg on IE)', () => {
-    const svgElement = appendElement('<svg class="foo bar"></svg>')
-    Object.defineProperty(svgElement, 'classList', { get: () => undefined })
-    const classList = getClassList(svgElement)
-    expect(classList[0]).toEqual('foo')
-    expect(classList[1]).toEqual('bar')
-  })
-})
-
-describe('getClassList', () => {
-  it('should return the classList of an element that supports the classList attribute', () => {
-    const divElement = appendElement('<div class="foo bar"></div>')
-    const classList = getClassList(divElement)
-    expect(classList[0]).toEqual('foo')
-    expect(classList[1]).toEqual('bar')
-  })
-
-  it('should return the classList of an element that does not support the classList attribute (e.g.: svg on IE)', () => {
-    const svgElement = appendElement('<svg class="foo bar"></svg>')
-    Object.defineProperty(svgElement, 'classList', { get: () => undefined })
-    const classList = getClassList(svgElement)
-    expect(classList[0]).toEqual('foo')
-    expect(classList[1]).toEqual('bar')
-  })
-})
