+++ packages/rum-core/src/index.ts
@@ -33,4 +33,5 @@ export { RumInitConfiguration, RumConfiguration } from './domain/configuration'
 export { DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from './domain/action/getActionNameFromElement'
 export { STABLE_ATTRIBUTES } from './domain/getSelectorFromElement'
 export * from './browser/htmlDomUtils'
+export * from './browser/polyfills'
 export { getSessionReplayUrl } from './domain/getSessionReplayUrl'
